+++ src/DateTime/tests/test_datetime.py
@@ -259,6 +259,13 @@ class DateTimeTests(unittest.TestCase):
         for key in DateTime.__slots__:
             self.assertEqual(getattr(dt, key), getattr(new, key))
 
+    def test_pickle_asdatetime_with_tz(self):
+        dt = DateTime('2002/5/2 8:00am GMT+8')
+        data = pickle.dumps(dt.asdatetime(), 1)
+        new = DateTime(pickle.loads(data))
+        for key in DateTime.__slots__:
+            self.assertEqual(getattr(dt, key), getattr(new, key))
+
     def test_pickle_with_numerical_tz(self):
         for dt_str in ('2007/01/02 12:34:56.789 +0300',
                        '2007/01/02 12:34:56.789 +0430',